FIX/REFACTOR: Reworked data to save/load to file, one dataset open at a time.
Changed data handling so that one dataset is loaded at a time, models are fitted to the dataset,
and then a new dataset is loaded. R doesn't handle dataframes in lists, so loading only one
dataset is much simpler. Data is now read from files in directory ./datasets/ and there is
a utility function save_builtin_datasets_to_file to populate the directory with builtin data.