REFACTOR/TEST: pass data as list-of-lists, do k-fold and predictor inclusion separately.
Rationale: nested iterations needed to generate lm data.
Specifically, need to iterate over datasets, k-folds, and parameters.
The easiest and most computationally efficient way to order the
iterations is datasets first (reduce disk access time), then
k-folds and parameters independently, so that the predictor graph
can be built independent of the fold. This refactor orders the
iterations in this order.

To unify the data type being passed from function to function,
I settled on a named list of lists - reason being data frames
don't work well with lm objects which are a type of named list.
Utility functions added to easily combine two lists of lists,
and by convention an empty list of lists is set to NULL so that
the details of what fields are in each list only need to be known
by the function which generates those fields - everything else
is just concatenated by append() if combining column-wise, or
append_list_of_lists() if combining row-wise.

To support future additions, code is now in ./src/, and tests
are in ./test/ although no particular test framework is being
used. NOTE: all files currently assume the working directory
is the parent directory, so run from src/data.R.