ADD: split datasets into smaller subsets using num_subsets parameter.
Changed k_fold_split to also split data into smaller subsets, allowing
models to be trained on less data than the dataset contains. This will
hopefully let the meta-model get more information about the effect of
number of observations on MSE.

There is a bug with powerTransform which breaks on some datasets when
the number of observations is too low. This was observed with Seatbelts.csv
when observations were less than 170. However, it didn't seem to affect
iris.csv even with as few as 40 observations. Cause is unknown.

Also need to update k_fold_split to still return useful data when
num_folds <= 1. This will make it easier to do the meta-model process.

Added model.R and updated main.R with function signatures for the meta-
model parameter selection process. None of this is working yet.