ADD: fit_model_to_predictors in model.R.
Added function to use meta model to choose predictors given a dataset.
Updated k_fold_split to return the entire dataset when num_folds=1.
Added parameter to load_dataframe to choose whether to normalize the
dataframe or not. Made graph.R return a directed graph so that the
model selection can choose forward/backward direction. Updated convenience
functions in graph.R to make it easier to get relevant vertices for
forward/backward steps, and changed the vertex to string conversion
so that it outputs T and F instead of 0 and 1, to avoid confusion with
numeric datatypes. Updated main.R to run fit_model_to_predictors.